+++ ubcpdk/simulations/circuits/test_sp.py
@@ -3,10 +3,10 @@ import pytest
 
 from gdslib.simphony.circuit import get_transmission
 
-from ubcsp import circuit_factory
-from ubcsp import component_factory
-from ubcsp import circuit_names
-from ubcsp import component_names
+from ubcpdk.simulations.circuits import circuit_factory
+from ubcpdk.simulations.circuits import component_factory
+from ubcpdk.simulations.circuits import circuit_names
+from ubcpdk.simulations.circuits import component_names
 
 
 @pytest.mark.parametrize("component_type", component_names)