+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.0.0
 
 - rename package from ubc to ubcpdk to match pypi name
-- move ubcsp into ubcpdk/simulations/circuits
+- move ubcpdk.simulations.circuits into ubcpdk/simulations/circuits
 - rename ubcpdk/da as ubcpdk/data
 
 ## 0.0.12